requirements.txt
mcp>=1.2.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
requests>=2.31.0
pydantic>=2.5.0
python-dotenv>=1.0.0